requirements.txt
streamlit
langchain
langchain-core
langchain-community
langchain-openai
faiss-cpu
python-dotenv
openai